@@ -2,13 +2,11 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
+using System.Diagnostics;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.Azure.Devices.Common.Exceptions;
 using Microsoft.Azure.Devices.E2ETests.Helpers;
-using Microsoft.Azure.Storage.Blob;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
 
@@ -33,19 +31,9 @@ namespace Microsoft.Azure.Devices.E2ETests.Iothub.Service
             '\n',
         };
 
-        private static readonly IReadOnlyList<JobStatus> s_incompleteJobs = new[]
-        {
-            JobStatus.Running,
-            JobStatus.Enqueued,
-            JobStatus.Queued,
-            JobStatus.Scheduled,
-            JobStatus.Unknown,
-        };
-
         [LoggedTestMethod, Timeout(LongRunningTestTimeoutMilliseconds)]
         [TestCategory("LongRunning")]
         [DoNotParallelize] // the number of jobs that can be run at a time are limited anyway
-        [Ignore]
         [DataRow(StorageAuthenticationType.KeyBased, false)]
         [DataRow(StorageAuthenticationType.IdentityBased, false)]
         [DataRow(StorageAuthenticationType.IdentityBased, true)]
@@ -126,7 +114,7 @@ namespace Microsoft.Azure.Devices.E2ETests.Iothub.Service
 
                 // act
 
-                JobProperties exportJobResponse = await CreateAndWaitForJobAsync(
+                JobProperties jobProperties = await CreateAndWaitForJobAsync(
                         storageAuthenticationType,
                         isUserAssignedMsi,
                         devicesFileName,
@@ -136,13 +124,17 @@ namespace Microsoft.Azure.Devices.E2ETests.Iothub.Service
                     .ConfigureAwait(false);
 
                 // assert
+
+                jobProperties.Status.Should().Be(JobStatus.Completed, $"Export failed due to '{jobProperties.FailureReason}'.");
+                jobProperties.FailureReason.Should().BeNullOrEmpty("Otherwise export failed");
+
                 await ValidateDevicesAsync(
                         devicesFileName,
                         storageContainer,
                         edge1,
                         edge2,
                         device)
-                .ConfigureAwait(false);
+                    .ConfigureAwait(false);
                 await ValidateConfigurationsAsync(
                         configsFileName,
                         storageContainer,
@@ -163,8 +155,6 @@ namespace Microsoft.Azure.Devices.E2ETests.Iothub.Service
             RegistryManager registryManager,
             Uri containerUri)
         {
-            int tryCount = 0;
-
             ManagedIdentity identity = isUserAssignedMsi
                 ? new ManagedIdentity
                 {
@@ -172,46 +162,50 @@ namespace Microsoft.Azure.Devices.E2ETests.Iothub.Service
                 }
                 : null;
 
-            JobProperties exportJobResponse = JobProperties.CreateForExportJob(
+            JobProperties jobProperties = JobProperties.CreateForExportJob(
                 containerUri.ToString(),
                 true,
                 devicesFileName,
                 storageAuthenticationType,
                 identity);
-            exportJobResponse.IncludeConfigurations = true;
-            exportJobResponse.ConfigurationsBlobName = configsFileName;
+            jobProperties.IncludeConfigurations = true;
+            jobProperties.ConfigurationsBlobName = configsFileName;
 
-            while (tryCount < MaxIterationWait)
+            var sw = Stopwatch.StartNew();
+
+            while (!jobProperties.IsFinished)
             {
                 try
                 {
-                    exportJobResponse = await registryManager.ExportDevicesAsync(exportJobResponse).ConfigureAwait(false);
+                    jobProperties = await registryManager.ExportDevicesAsync(jobProperties).ConfigureAwait(false);
+                    if (!string.IsNullOrWhiteSpace(jobProperties.FailureReason))
+                    {
+                        Logger.Trace($"Job failed due to {jobProperties.FailureReason}");
+                    }
                     break;
                 }
                 // Concurrent jobs can be rejected, so implement a retry mechanism to handle conflicts with other tests
-                catch (JobQuotaExceededException) when (++tryCount < MaxIterationWait)
+                catch (JobQuotaExceededException)
                 {
-                    Logger.Trace($"JobQuotaExceededException... waiting.");
+                    Logger.Trace($"JobQuotaExceededException... waiting after {sw.Elapsed}.");
                     await Task.Delay(s_waitDuration).ConfigureAwait(false);
                     continue;
                 }
             }
 
-            for (int i = 0; i < MaxIterationWait; ++i)
+            sw.Stop();
+            Logger.Trace($"Job started after {sw.Elapsed}.");
+
+            sw.Restart();
+
+            while (!jobProperties.IsFinished)
             {
                 await Task.Delay(s_waitDuration).ConfigureAwait(false);
-                exportJobResponse = await registryManager.GetJobAsync(exportJobResponse.JobId).ConfigureAwait(false);
-                Logger.Trace($"Job {exportJobResponse.JobId} is {exportJobResponse.Status} with progress {exportJobResponse.Progress}%");
-                if (!s_incompleteJobs.Contains(exportJobResponse.Status))
-                {
-                    break;
-                }
+                jobProperties = await registryManager.GetJobAsync(jobProperties.JobId).ConfigureAwait(false);
+                Logger.Trace($"Job {jobProperties.JobId} is {jobProperties.Status} with progress {jobProperties.Progress}% after {sw.Elapsed}.");
             }
 
-            exportJobResponse.Status.Should().Be(JobStatus.Completed, "Otherwise import failed");
-            exportJobResponse.FailureReason.Should().BeNullOrEmpty("Otherwise import failed");
-
-            return exportJobResponse;
+            return jobProperties;
         }
 
         private async Task ValidateDevicesAsync(
@@ -299,7 +293,7 @@ namespace Microsoft.Azure.Devices.E2ETests.Iothub.Service
 
         private static async Task<string> DownloadFileAsync(StorageContainer storageContainer, string fileName)
         {
-            CloudBlockBlob exportFile = storageContainer.CloudBlobContainer.GetBlockBlobReference(fileName);
+            Storage.Blob.CloudBlockBlob exportFile = storageContainer.CloudBlobContainer.GetBlockBlobReference(fileName);
             return await exportFile.DownloadTextAsync().ConfigureAwait(false);
         }
 
@@ -313,13 +307,37 @@ namespace Microsoft.Azure.Devices.E2ETests.Iothub.Service
             try
             {
                 await registryManager.RemoveDeviceAsync(deviceId).ConfigureAwait(false);
+            }
+            catch (Exception ex)
+            {
+                Logger.Trace($"Failed to remove device during cleanup due to {ex.Message}");
+            }
+
+            try
+            {
                 await registryManager.RemoveDeviceAsync(edgeId2).ConfigureAwait(false);
+            }
+            catch (Exception ex)
+            {
+                Logger.Trace($"Failed to remove device during cleanup due to {ex.Message}");
+            }
+
+            try
+            {
                 await registryManager.RemoveDeviceAsync(edgeId1).ConfigureAwait(false);
+            }
+            catch (Exception ex)
+            {
+                Logger.Trace($"Failed to remove device during cleanup due to {ex.Message}");
+            }
+
+            try
+            {
                 await registryManager.RemoveConfigurationAsync(configurationId).ConfigureAwait(false);
             }
             catch (Exception ex)
             {
-                Logger.Trace($"Failed to remove device/config during cleanup due to {ex}");
+                Logger.Trace($"Failed to remove config during cleanup due to {ex.Message}");
             }
         }
     }
